cosas

+++ add.h
@@ -1,4 +1,4 @@
 #include "types.h"
 #include "library.h"
 
-int add(Index *index, long book_id, char *isbn, char *title, char *printedBy);
+int add(Index *index, int book_id, char *isbn, char *title, char *printedBy);
